pyproject.toml
[build-system]
requires = ["poetry-core>=1,<2", "poetry-dynamic-versioning>=1.4,<2"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
name = "pychip8"
version = "0.0.0"
description = "Chip-8 emulator written in Python"
authors = ["Eduardo Klosowski <eduardo_klosowski@yahoo.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/eduardoklosowski/pychip8"
classifiers = [
  "Development Status :: 1 - Planning",
  "Environment :: Console :: Curses",
  "Environment :: X11 Applications",
  "Intended Audience :: End Users/Desktop",
  "Operating System :: OS Independent",
  "Topic :: Games/Entertainment",
  "Topic :: Software Development :: Interpreters",
  "Topic :: System :: Emulators",
  "Typing :: Typed",
  "Private :: Do Not Upload",
]

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"

[tool.poetry.dependencies]
python = "^3.10"
pysdl2 = "^0.9"

[tool.poetry.group.lint.dependencies]
ruff = "^0.6"
mypy = "^1.11"

[tool.poetry.group.test.dependencies]
pytest = "^8.3"
pytest-cov = "^5.0"

[tool.poetry.scripts]
pychip8 = "pychip8.cli:main"

[tool.ruff]
target-version = "py310"
line-length = 119
src = ["src", "tests"]

[tool.ruff.format]
line-ending = "lf"
quote-style = "single"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "FBT", "COM812", "EM", "ISC001", "PLR0913", "TRY003"]

[tool.ruff.lint.per-file-ignores]
"src/pychip8/core.py" = ["PLR2004"]
"tests/*.py" = ["S101", "S311", "SLF001", "PLR2004"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.mypy]
warn_unused_configs = true
show_column_numbers = true
show_error_codes = true
sqlite_cache = true
strict = true
files = ["src/**/*.py", "tests/**/*.py"]

[[tool.mypy.overrides]]
module = [
  "sdl2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if (typing\\.)?TYPE_CHECKING:",
  "@overload",
  "class .*\\((typing\\.)?Protocol\\):",
  "if __name__ == '__main__':",
]
